+++ src/TextureManager.cpp
@@ -6,11 +6,14 @@
  * \author xythobuz
  */
 
+#include "imgui/imgui.h"
 #include "stb/stb_image.h"
 
 #include "global.h"
+#include "Game.h"
 #include "Log.h"
 #include "RunTime.h"
+#include "World.h"
 #include "utils/Folder.h"
 #include "utils/pcx.h"
 #include "utils/pixel.h"
@@ -228,15 +231,14 @@ int TextureManager::getFirstTileAnimation(int index) {
     return animations.at(index).at(0);
 }
 
-int TextureManager::getNextTileAnimation(int tile) {
-    for (int a = 0; a < animations.size(); a++) {
-        for (int i = 0; i < animations.at(a).size(); i++) {
-            if (animations.at(a).at(i) == tile) {
-                if (i < (animations.at(a).size() - 1))
-                    return animations.at(a).at(i + 1);
-                else
-                    return animations.at(a).at(0);
-            }
+int TextureManager::getNextTileAnimation(int index, int tile) {
+    assert(index < animations.size());
+    for (int i = 0; i < animations.at(index).size(); i++) {
+        if (animations.at(index).at(i) == tile) {
+            if (i < (animations.at(index).size() - 1))
+                return animations.at(index).at(i + 1);
+            else
+                return animations.at(index).at(0);
         }
     }
     return -1;
@@ -311,6 +313,206 @@ std::vector<int>& TextureManager::getUnits(TextureStorage s) {
         return gameUnits;
     else
         return systemUnits;
+}
+
+void TextureManager::display() {
+    if (ImGui::CollapsingHeader("Texture Viewer")) {
+        static bool game = Game::isLoaded();
+        static int index = 0;
+        ImGui::SliderInt("##texslide", &index, 0, TextureManager::numTextures(
+                             game ? TextureStorage::GAME : TextureStorage::SYSTEM) - 1);
+        ImGui::SameLine();
+        if (ImGui::Button("+##texplus", ImVec2(0, 0), true)) {
+            if (index < (numTextures(
+                             game ? TextureStorage::GAME : TextureStorage::SYSTEM) - 1))
+                index++;
+            else
+                index = 0;
+        }
+        ImGui::SameLine();
+        if (ImGui::Button("-##texminus", ImVec2(0, 0), true)) {
+            if (index > 0)
+                index--;
+            else
+                index = numTextures(
+                            game ? TextureStorage::GAME : TextureStorage::SYSTEM) - 1;
+        }
+
+        if ((numTextures(TextureStorage::GAME) > 0)) {
+            ImGui::SameLine();
+            ImGui::Checkbox("Game##texgame", &game);
+        } else {
+            game = false;
+        }
+
+        if (index >= numTextures(game ? TextureStorage::GAME : TextureStorage::SYSTEM)) {
+            index = numTextures(game ? TextureStorage::GAME : TextureStorage::SYSTEM) - 1;
+            if (index < 0) {
+                game = false;
+                index = 0;
+            }
+        }
+
+        auto bm = getBufferManager(index, game ? TextureStorage::GAME
+                                                            : TextureStorage::SYSTEM);
+        ImGui::Image(bm, ImVec2(ImGui::GetColumnWidth() * 2 / 3, ImGui::GetColumnWidth() * 2 / 3));
+    }
+
+    if (ImGui::CollapsingHeader("Textile Viewer")) {
+        if (numTiles() > 0) {
+            static int index = 0;
+            ImGui::PushItemWidth(ImGui::GetWindowWidth() * 0.5f);
+            ImGui::SliderInt("##tileslide", &index, 0, numTiles() - 1);
+            ImGui::PopItemWidth();
+            ImGui::SameLine();
+            if (ImGui::Button("+##tileplus", ImVec2(0, 0), true)) {
+                if (index < (numTiles() - 1))
+                    index++;
+                else
+                    index = 0;
+            }
+            ImGui::SameLine();
+            if (ImGui::Button("-##tileminus", ImVec2(0, 0), true)) {
+                if (index > 0)
+                    index--;
+                else
+                    index = numTiles() - 1;
+            }
+
+            if (index >= numTiles())
+                index = 0;
+
+            auto& tile = getTile(index);
+            auto bm = getBufferManager(tile.getTexture(), TextureStorage::GAME);
+            ImVec2 size(ImGui::GetColumnWidth() * 2 / 3, ImGui::GetColumnWidth() * 2 / 3);
+            auto uvA = tile.getUV(0);
+            auto uvB = tile.getUV(2);
+            ImVec2 uv1(uvA.x, uvA.y);
+            ImVec2 uv2(uvB.x, uvB.y);
+            ImGui::Image(bm, size, uv1, uv2);
+        } else {
+            ImGui::Text("No textiles are currently loaded...!");
+        }
+    }
+
+    if (ImGui::CollapsingHeader("Animated Textile Viewer")) {
+        if (numAnimatedTiles() > 0) {
+            static int index = 0;
+            static int tile = getFirstTileAnimation(index);
+            if (ImGui::SliderInt("##animslide", &index, 0, numAnimatedTiles() - 1)) {
+                tile = getFirstTileAnimation(index);
+            }
+            ImGui::SameLine();
+            if (ImGui::Button("+##animplus", ImVec2(0, 0), true)) {
+                if (index < (numAnimatedTiles() - 1))
+                    index++;
+                else
+                    index = 0;
+                tile = getFirstTileAnimation(index);
+            }
+            ImGui::SameLine();
+            if (ImGui::Button("-##animminus", ImVec2(0, 0), true)) {
+                if (index > 0)
+                    index--;
+                else
+                    index = numAnimatedTiles() - 1;
+                tile = getFirstTileAnimation(index);
+            }
+
+            if (index >= numAnimatedTiles()) {
+                index = 0;
+                tile = getFirstTileAnimation(index);
+            }
+
+            int next = getNextTileAnimation(index, tile);
+            if (next == -1) {
+                index = 0;
+                tile = getFirstTileAnimation(index);
+            }
+
+            ImGui::SameLine();
+            ImGui::Text("%d", tile);
+
+            auto& t = getTile(tile);
+            auto bm = getBufferManager(t.getTexture(), TextureStorage::GAME);
+            ImVec2 size(ImGui::GetColumnWidth() * 2 / 3, ImGui::GetColumnWidth() * 2 / 3);
+            auto uvA = t.getUV(0);
+            auto uvB = t.getUV(2);
+            ImVec2 uv1(uvA.x, uvA.y);
+            ImVec2 uv2(uvB.x, uvB.y);
+            ImGui::Image(bm, size, uv1, uv2);
+
+            static int fr = 0;
+            if (fr > 0) {
+                fr--;
+            } else {
+                fr = RunTime::getFPS() / 2;
+                tile = next;
+            }
+        } else {
+            ImGui::Text("No animated textures are currently loaded...!");
+        }
+    }
+
+    if (ImGui::CollapsingHeader("Sprite Sequence Viewer")) {
+        if (getWorld().sizeSprite() <= 0) {
+            ImGui::Text("Please load a level containing sprites!");
+        } else {
+            static int index = 0;
+            static int sprite = 0;
+            if (ImGui::SliderInt("##spriteslide", &index, 0, getWorld().sizeSpriteSequence() - 1)) {
+                sprite = 0;
+            }
+            ImGui::SameLine();
+            if (ImGui::Button("+##spriteplus", ImVec2(0, 0), true)) {
+                if (index < (getWorld().sizeSpriteSequence() - 1))
+                    index++;
+                else
+                    index = 0;
+                sprite = 0;
+            }
+            ImGui::SameLine();
+            if (ImGui::Button("-##spriteminus", ImVec2(0, 0), true)) {
+                if (index > 0)
+                    index--;
+                else
+                    index = getWorld().sizeSpriteSequence() - 1;
+                sprite = 0;
+            }
+
+            if (index >= getWorld().sizeSpriteSequence()) {
+                index = 0;
+                sprite = 0;
+            }
+
+            if (sprite >= getWorld().getSpriteSequence(index).size()) {
+                sprite = 0;
+            }
+
+            ImGui::SameLine();
+            ImGui::Text("Sprite %d/%d", sprite + 1, getWorld().getSpriteSequence(index).size());
+
+            auto& s = getWorld().getSprite(getWorld().getSpriteSequence(index).getStart() + sprite);
+            auto bm = getBufferManager(s.getTexture(), TextureStorage::GAME);
+            ImVec2 size(ImGui::GetColumnWidth() * 2 / 3, ImGui::GetColumnWidth() * 2 / 3);
+            auto uv = s.getUVs();
+            ImVec2 uv1(uv.x, uv.w);
+            ImVec2 uv2(uv.z, uv.y);
+            ImGui::Image(bm, size, uv1, uv2);
+
+            static int fr = 0;
+            if (fr > 0) {
+                fr--;
+            } else {
+                fr = RunTime::getFPS() / 10;
+                if (sprite < (getWorld().getSpriteSequence(index).size() - 1))
+                    sprite++;
+                else
+                    sprite = 0;
+            }
+
+        }
+    }
 }
 
 // ----------------------------------------------------------------------------